--- libs/game---hw/pxt.json
@@ -1,10 +1,0 @@
-{
-    "name": "game",
-    "additionalFilePath": "../game",
-    "partial": true,
-    "dependencies": {
-        "settings": "file:../settings",
-        "screen---st7735": "file:../screen---st7735",
-        "power": "file:../power"
-    }
-}
